+++ private/test/src/setup.ts
@@ -1,1 +1,0 @@
-import "@ungap/with-resolvers"; // Node.js 20 support